tests/wpt/tests/permissions-policy/reporting/encrypted-media-report-only.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
var check_report_format = ([reports, observer]) => {
  let report = reports[0];
  assert_equals(report.type, "permissions-policy-violation");
  assert_equals(report.body.featureId, "encrypted-media");
  assert_equals(report.body.disposition, "report");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['permissions-policy-violation']}).observe();
  });
  await navigator.requestMediaKeySystemAccess("org.w3.clearkey",
      [{
        initDataTypes: ["webm"],
        videoCapabilities: [{contentType: 'video/webm;codecs="vp8"'}],
      }]);
  check_report_format(await report);
}, "Encrypted Media report only mode");
    </script>
  </body>
</html>
